% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_model.R
\name{dynamic_impact_vector_variate}
\alias{dynamic_impact_vector_variate}
\title{dynamic_impact_vector_variate
Generate specific summaries of numeric columns in a data frame}
\usage{
dynamic_impact_vector_variate(
  Y,
  X,
  initial_intervention_index,
  keep_evolution_static_for_prediction = TRUE,
  share_columns_var = FALSE,
  final_intervention_date = NULL,
  predefined_variance = NULL,
  variable_names = NULL,
  return_stan_output = FALSE,
  return_df_format = TRUE,
  confidence_level = 0.9,
  time_index = NULL,
  n_cores = NULL,
  ...
)
}
\description{
dynamic_impact_vector_variate
Generate specific summaries of numeric columns in a data frame
}
