% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicimpact-package.R
\docType{package}
\name{dynamicimpact-package}
\alias{dynamicimpact-package}
\alias{dynamicimpact}
\title{Inferring Causal Effects on Vector and Matrix Variate Time Series Using Bayesian Dynamic Models.}
\description{
The functionality of this software is to determine if there is a significant impact on a system after, a event of interest with known start date \code{T_0}. This package supports vector and matrix variate time series.
}
\section{Authors}{

Maintainer: Carlos Peréz \email{caaperezag@unal.edu.co} \cr
Creator: Johnatan Cardona-Jiménez \email{jcardonj@unal.edu.co} \cr
Creator: Isabel Ramírez \email{iscramirezgu@unal.edu.co}
}

\references{
Stan Development Team (2023). RStan: the R interface to Stan. R package version 2.21.8. https://mc-stan.org
}
