% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot.impact_vector_model_cholesky}
\alias{plot.impact_vector_model_cholesky}
\title{plot.impact_vector_model_cholesky}
\usage{
\method{plot}{impact_vector_model_cholesky}(x, plot_variable = "global", break_dates = NULL, ...)
}
\arguments{
\item{x:}{fitted model.}

\item{plot_variable:}{variable to be plotted use \code{"global"} for the aggregated result}

\item{break_dates:}{vector of breaks for the x axis.}
}
\value{
A \code{ggplot2} object
}
\description{
plot a fitted model.
}
\details{
plots \code{ImpactModelVector} object.
}
\examples{
\donttest{
data("X_vector", package="dynamicimpact")
data("Y_vector", package="dynamicimpact")
result <- ImpactModelVector(X_data=X_vector,
                            Y_data=Y_vector, event_initial=97,
                            credibility_level=0.89, use_cholesky=TRUE)
plot(result)
}
}
