% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot.model_impact_vector}
\alias{plot.model_impact_vector}
\title{plot.model_impact_vector}
\usage{
\method{plot}{model_impact_vector}(x, plot_variable = "global", break_dates = NULL, ...)
}
\arguments{
\item{x:}{fitted model.}

\item{plot_variable:}{variable to be plotted use \code{"global"} for the aggregated result}

\item{break_dates:}{vector of breaks for the x axis.}
}
\value{
A \code{ggplot2} object
}
\description{
plot.model_impact_vector
}
\details{
plots \code(ImpactModelVector) object.
}
\examples{
\donttest{
data("X_vector", package="dynamicimpact")
data("Y_vector", package="dynamicimpact")
result <- ImpactModelVector(X_data=X_vector, Y_data=Y_vector, event_initial=97, credibility_level=0.89)
plot(result)
}
}
